Rename raids that are listed as dungeons that are actually raids.
