+++ src/scripts/InstanceScripts/Wotlk/Raid_VaultOfArchavon.cpp
@@ -4,7 +4,7 @@ This file is released under the MIT license. See README-MIT for more information
 */
 
 #include "Setup.h"
-#include "Instance_VaultOfArchavon.h"
+#include "Raid_VaultOfArchavon.h"
 #include <Units/Creatures/Pet.h>
 #include "Server/Script/CreatureAIScript.h"
