+++ src/scripts/InstanceScripts/Wotlk/Raid_VaultOfArchavon.h
